Updated DailyScrum, ScrumsView and Theme. Created DetailView
On DailyScrum: an attendees struct has been created inside of an extension. Now attendees complies with the Identifiable protocol.

On ScrumsView: now is inside of a NavigationStack that shows DetailView

Theme: the name of the theme was capitalised thought a computed variable

DetailView: it shows the details of the scrums. It has a button that will navigate to the MeetingView